Consistent naming; cdc is same client/server.
