+++ cadence/transactions/kittyItemsMarket/buy_market_item.cdc
@@ -4,28 +4,24 @@ import Kibble from "../../contracts/Kibble.cdc"
 import KittyItems from "../../contracts/KittyItems.cdc"
 import KittyItemsMarket from "../../contracts/KittyItemsMarket.cdc"
 
+
 transaction(itemID: UInt64, marketCollectionAddress: Address) {
     let paymentVault: @FungibleToken.Vault
     let kittyItemsCollection: &KittyItems.Collection{NonFungibleToken.Receiver}
     let marketCollection: &KittyItemsMarket.Collection{KittyItemsMarket.CollectionPublic}
 
-    prepare(signer: AuthAccount) {
+    prepare(acct: AuthAccount) {
         self.marketCollection = getAccount(marketCollectionAddress)
-            .getCapability<&KittyItemsMarket.Collection{KittyItemsMarket.CollectionPublic}>(
-                KittyItemsMarket.CollectionPublicPath
-            )!
-            .borrow()
-            ?? panic("Could not borrow market collection from market address")
+            .getCapability<&KittyItemsMarket.Collection{KittyItemsMarket.CollectionPublic}>(KittyItemsMarket.CollectionPublicPath)
+            .borrow() ?? panic("Could not borrow market collection from market address")
 
-        let saleItem = self.marketCollection.borrowSaleItem(itemID: itemID)
-                    ?? panic("No item with that ID")
-        let price = saleItem.price
+        let price = self.marketCollection.borrowSaleItem(itemID: itemID)!.price
 
-        let mainKibbleVault = signer.borrow<&Kibble.Vault>(from: Kibble.VaultStoragePath)
+        let mainKibbleVault = acct.borrow<&Kibble.Vault>(from: Kibble.VaultStoragePath)
             ?? panic("Cannot borrow Kibble vault from acct storage")
         self.paymentVault <- mainKibbleVault.withdraw(amount: price)
 
-        self.kittyItemsCollection = signer.borrow<&KittyItems.Collection{NonFungibleToken.Receiver}>(
+        self.kittyItemsCollection = acct.borrow<&KittyItems.Collection{NonFungibleToken.Receiver}>(
             from: KittyItems.CollectionStoragePath
         ) ?? panic("Cannot borrow KittyItems collection receiver from acct")
     }
