+++ web/src/flow/script.is-account-initialized.js
