+++ web/src/flow/tx.buy-market-item.js
@@ -1,0 +1,26 @@
+import * as fcl from "@onflow/fcl"
+import * as t from "@onflow/types"
+import {tx} from "./util/tx"
+import {invariant} from "@onflow/util-invariant"
+
+const CODE = fcl.cdc`
+  // CADENCE TODO
+`
+
+// prettier-ignore
+export function buyMarketItem({itemID, ownerAddress}, opts = {}) {
+  invariant(itemID != null, "buyMarketItem({itemID, ownerAddress}) -- itemID required")
+  invariant(ownerAddress != null, "buyMarketItem({itemID, ownerAddress}) -- ownerAddress required")
+
+  return tx([
+    fcl.transaction(CODE),
+    fcl.args([
+      fcl.arg(Number(itemID), t.UInt64),
+      fcl.arg(String(ownerAddress), t.Address),
+    ]),
+    fcl.proposer(fcl.authz),
+    fcl.payer(fcl.authz),
+    fcl.authorizations([fcl.authz]),
+    fcl.limit(1000),
+  ], opts)
+}
