+++ web/src/flow/tx.remove-sale-offer.js
@@ -6,15 +6,19 @@ import {invariant} from "@onflow/util-invariant"
 const CODE = fcl.cdc`
   import KittyItemsMarket from 0xKittyItemsMarket
 
-  transaction(itemID: UInt64) {
-      prepare(account: AuthAccount) {
-          let listing <- account
-            .borrow<&KittyItemsMarket.Collection>(from: KittyItemsMarket.CollectionStoragePath)!
-            .remove(itemID: itemID)
-          destroy listing
-      }
-  }
-`
+transaction(itemID: UInt64) {
+    let marketCollection: &KittyItemsMarket.Collection
+
+    prepare(signer: AuthAccount) {
+        self.marketCollection = signer.borrow<&KittyItemsMarket.Collection>(from: KittyItemsMarket.CollectionStoragePath)
+            ?? panic("Missing or mis-typed KittyItemsMarket Collection")
+    }
+
+    execute {
+        let offer <-self.marketCollection.remove(itemID: itemID)
+        destroy offer
+    }
+}`
 
 // prettier-ignore
 export function cancelMarketListing({ itemID }, opts = {}) {
