+++ web/src/util/cdc-loader.js
@@ -1,0 +1,5 @@
+import raw from "raw.macro"
+
+export const loadTransaction = () => {}
+
+export const loadScript = () => {}
